Refractor occupancy into seperate berth module
