+++ crates/dock-alloc-model/src/lib.rs
@@ -754,53 +754,6 @@ where
     }
 }
 
-impl<T: PrimInt + Signed, C: PrimInt + Signed> From<Assignment<Movable, T, C>>
-    for AnyAssignment<T, C>
-{
-    fn from(a: Assignment<Movable, T, C>) -> Self {
-        AnyAssignment::new(
-            AnyRequest::Movable(a.request().clone()),
-            a.start_position(),
-            a.start_time(),
-        )
-    }
-}
-impl<T: PrimInt + Signed, C: PrimInt + Signed> From<Assignment<Fixed, T, C>>
-    for AnyAssignment<T, C>
-{
-    fn from(a: Assignment<Fixed, T, C>) -> Self {
-        AnyAssignment::new(
-            AnyRequest::Fixed(a.request().clone()),
-            a.start_position(),
-            a.start_time(),
-        )
-    }
-}
-
-// From AssignmentRef (Copy) → owned erased view
-impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<AssignmentRef<'a, Movable, T, C>>
-    for AnyAssignment<T, C>
-{
-    fn from(a: AssignmentRef<'a, Movable, T, C>) -> Self {
-        AnyAssignment::new(
-            AnyRequest::Movable(a.request().clone()),
-            a.start_position(),
-            a.start_time(),
-        )
-    }
-}
-impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<AssignmentRef<'a, Fixed, T, C>>
-    for AnyAssignment<T, C>
-{
-    fn from(a: AssignmentRef<'a, Fixed, T, C>) -> Self {
-        AnyAssignment::new(
-            AnyRequest::Fixed(a.request().clone()),
-            a.start_position(),
-            a.start_time(),
-        )
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct AnyAssignmentRef<'a, T = i64, C = i64>
 where
@@ -866,10 +819,10 @@ where
     }
 }
 
-// From owned Assignment → borrowed erased view (we borrow its inner request)
 impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&'a Assignment<Movable, T, C>>
     for AnyAssignmentRef<'a, T, C>
 {
+    #[inline]
     fn from(a: &'a Assignment<Movable, T, C>) -> Self {
         AnyAssignmentRef::new(
             AnyRequestRef::Movable(a.request()),
@@ -878,12 +831,140 @@ impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&'a Assignment<Movable, 
         )
     }
 }
+
 impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&'a Assignment<Fixed, T, C>>
     for AnyAssignmentRef<'a, T, C>
 {
+    #[inline]
     fn from(a: &'a Assignment<Fixed, T, C>) -> Self {
         AnyAssignmentRef::new(
             AnyRequestRef::Fixed(a.request()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<T: PrimInt + Signed, C: PrimInt + Signed> From<&Assignment<Movable, T, C>>
+    for AnyAssignment<T, C>
+{
+    #[inline]
+    fn from(a: &Assignment<Movable, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Movable(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<T: PrimInt + Signed, C: PrimInt + Signed> From<&Assignment<Fixed, T, C>>
+    for AnyAssignment<T, C>
+{
+    #[inline]
+    fn from(a: &Assignment<Fixed, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Fixed(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<T: PrimInt + Signed, C: PrimInt + Signed> From<Assignment<Movable, T, C>>
+    for AnyAssignment<T, C>
+{
+    fn from(a: Assignment<Movable, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Movable(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<T: PrimInt + Signed, C: PrimInt + Signed> From<Assignment<Fixed, T, C>>
+    for AnyAssignment<T, C>
+{
+    fn from(a: Assignment<Fixed, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Fixed(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&'a AssignmentRef<'a, Movable, T, C>>
+    for AnyAssignmentRef<'a, T, C>
+{
+    #[inline]
+    fn from(a: &'a AssignmentRef<'a, Movable, T, C>) -> Self {
+        AnyAssignmentRef::new(
+            AnyRequestRef::Movable(a.request()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&'a AssignmentRef<'a, Fixed, T, C>>
+    for AnyAssignmentRef<'a, T, C>
+{
+    #[inline]
+    fn from(a: &'a AssignmentRef<'a, Fixed, T, C>) -> Self {
+        AnyAssignmentRef::new(
+            AnyRequestRef::Fixed(a.request()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&AssignmentRef<'a, Movable, T, C>>
+    for AnyAssignment<T, C>
+{
+    #[inline]
+    fn from(a: &AssignmentRef<'a, Movable, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Movable(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<&AssignmentRef<'a, Fixed, T, C>>
+    for AnyAssignment<T, C>
+{
+    #[inline]
+    fn from(a: &AssignmentRef<'a, Fixed, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Fixed(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<AssignmentRef<'a, Movable, T, C>>
+    for AnyAssignment<T, C>
+{
+    fn from(a: AssignmentRef<'a, Movable, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Movable(a.request().clone()),
+            a.start_position(),
+            a.start_time(),
+        )
+    }
+}
+
+impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<AssignmentRef<'a, Fixed, T, C>>
+    for AnyAssignment<T, C>
+{
+    fn from(a: AssignmentRef<'a, Fixed, T, C>) -> Self {
+        AnyAssignment::new(
+            AnyRequest::Fixed(a.request().clone()),
             a.start_position(),
             a.start_time(),
         )
@@ -901,6 +982,7 @@ impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<AssignmentRef<'a, Movabl
         )
     }
 }
+
 impl<'a, T: PrimInt + Signed, C: PrimInt + Signed> From<AssignmentRef<'a, Fixed, T, C>>
     for AnyAssignmentRef<'a, T, C>
 {
