+++ crates/dock-alloc-solver/src/berth/mod.rs
@@ -19,8 +19,10 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-pub mod berth;
-pub mod container;
+pub mod berthocc;
 pub mod domain;
-pub mod framework;
-pub mod ledger;
+mod iter;
+pub mod operations;
+pub mod overlay;
+pub mod quay;
+pub mod slice;
