+++ crates/dock-alloc-solver/src/berth/quay.rs
