+++ crates/dock-alloc-solver/src/container/intervalset.rs
