+++ crates/dock-alloc-solver/src/container/timeline.rs
