+++ crates/dock-alloc-solver/src/framework.rs
@@ -19,9 +19,13 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-use crate::{ledger::AssignmentLedger, occupancy::BerthOccupancy, quay::QuayRead};
 use dock_alloc_model::{OwnedSolution, Problem, SolutionRef};
 use num_traits::{PrimInt, Signed};
+
+use crate::{
+    berth::{berthocc::BerthOccupancy, quay::QuayRead},
+    ledger::AssignmentLedger,
+};
 
 pub struct SearchState<'p, T, C, Q>
 where
